WIP add l90c and adj123 fields to 12.5km ancillary files
