+++ seaice_ecdr/constants.py
@@ -25,5 +25,3 @@ LANCE_NRT_DATA_DIR = NSIDC_NFS_SHARE_DIR / "lance_amsr2_nrt_data"
 
 # Location of surface mask & geo-information files.
 CDR_ANCILLARY_DIR = NSIDC_NFS_SHARE_DIR / "cdrv5_ancillary"
-
-NASATEAM2_ANCILLARY_DIR = NSIDC_NFS_SHARE_DIR / "nasateam2_ancillary"
