--- seaice_ecdr/land_spillover.py
@@ -1,78 +1,0 @@
-from pathlib import Path
-
-import numpy as np
-from loguru import logger
-from pm_icecon.land_spillover import create_land90
-
-from seaice_ecdr.constants import NASATEAM2_ANCILLARY_DIR
-
-# TODO: The various directory vars, eg anc_dir, should be either abstracted
-#   as a constant or passed as a configuration parameter.
-# TODO: So too, the filename template strings should be authoritatively
-#   set in a central location.
-
-
-def read_adj123_file(
-    grid_id: str = "psn12.5",
-    xdim: int = 608,
-    ydim: int = 896,
-    anc_dir: Path = NASATEAM2_ANCILLARY_DIR,
-    adj123_fn_template: str = "{anc_dir}/coastal_adj_diag123_{grid_id}.dat",
-):
-    """Read the diagonal adjacency 123 file."""
-    coast_adj_fn = adj123_fn_template.format(anc_dir=anc_dir, grid_id=grid_id)
-    assert Path(coast_adj_fn).is_file()
-    adj123 = np.fromfile(coast_adj_fn, dtype=np.uint8).reshape(ydim, xdim)
-
-    return adj123
-
-
-def create_land90_conc_file(
-    grid_id: str = "psn12.5",
-    xdim: int = 608,
-    ydim: int = 896,
-    anc_dir: Path = NASATEAM2_ANCILLARY_DIR,
-    adj123_fn_template: str = "{anc_dir}/coastal_adj_diag123_{grid_id}.dat",
-    write_l90c_file: bool = True,
-    l90c_fn_template: str = "{anc_dir}/land90_conc_{grid_id}.dat",
-):
-    """Create the land90-conc file.
-
-    The 'land90' array is a mock sea ice concentration array that is calculated
-    from the land mask.  It assumes that the mock concentration value will be
-    the average of a 7x7 array of local surface mask values centered on the
-    center pixel.  Water grid cells are considered to have a sea ice
-    concentration of zero.  Land grid cells are considered to have a sea ice
-    concentration of 90%.  The average of the 49 grid cells in the 7x7 array
-    yields the `land90` concentration value.
-    """
-    adj123 = read_adj123_file(grid_id, xdim, ydim, anc_dir, adj123_fn_template)
-    land90 = create_land90(adj123=adj123)
-
-    if write_l90c_file:
-        l90c_fn = l90c_fn_template.format(anc_dir=anc_dir, grid_id=grid_id)
-        land90.tofile(l90c_fn)
-        print(f"Wrote: {l90c_fn}\n  {land90.dtype}  {land90.shape}")
-
-    return land90
-
-
-def load_or_create_land90_conc(
-    grid_id: str = "psn12.5",
-    xdim: int = 608,
-    ydim: int = 896,
-    anc_dir: Path = NASATEAM2_ANCILLARY_DIR,
-    l90c_fn_template: str = "{anc_dir}/land90_conc_{grid_id}.dat",
-    overwrite: bool = False,
-):
-    # Attempt to load the land90_conc field, and if fail, create it
-    l90c_fn = l90c_fn_template.format(anc_dir=anc_dir, grid_id=grid_id)
-    if overwrite or not Path(l90c_fn).is_file():
-        data = create_land90_conc_file(
-            grid_id, xdim, ydim, anc_dir=anc_dir, l90c_fn_template=l90c_fn_template
-        )
-    else:
-        data = np.fromfile(l90c_fn, dtype=np.float32).reshape(ydim, xdim)
-        logger.info(f"Read NT2 land90 mask from:\n  {l90c_fn}")
-
-    return data
